Fixed blog_url not returning & rake setup_github_pages failing
When executing `rake setup_github_pages`, an error tells that `url` don't exist. The reference to that variable has been changed to a call to `blog_url`.

The function `blog_url` was not returning anything, failing to properly write the `_config.yml` file. It now does return what should have.